fix type conversion

+++ Common/commands.hpp
@@ -44,7 +44,7 @@ DEF_CMD(POP, 7, POP__ARGS, {
 })
 
 DEF_CMD(JMP, 8, JUMP_ARGS, {
-    cpu->ip = *((int*) (cpu->code + cpu->ip) - 1);
+    cpu->ip = (size_t) *((int*) (cpu->code + cpu->ip) - 1);
 })
 
 #define DEF_JMP_IF(name, cmd_num, oper)                                \
@@ -53,7 +53,7 @@ DEF_CMD(JMP, 8, JUMP_ARGS, {
             second_popped = StackPop (cpu->cpu_stack, &cpu->stk_err);  \
                                                                        \
             if (first__popped oper second_popped) {                    \
-                cpu->ip = *((int *) (cpu->code + cpu->ip) - 1);        \
+                cpu->ip = (size_t) *((int *) (cpu->code + cpu->ip) - 1);        \
             }                                                          \
         })
 
@@ -68,11 +68,11 @@ DEF_JMP_IF(JNE, 14, !=)
 
 DEF_CMD(CALL, 15, JUMP_ARGS, {
     StackPush(cpu->adr_stack, (int) cpu->ip);
-    cpu->ip = *((int*) (cpu->code + cpu->ip) - 1);
+    cpu->ip = (size_t) *((int*) (cpu->code + cpu->ip) - 1);
 })
 
 DEF_CMD(RET, 16, NO_ARGS, {
-    cpu->ip = StackPop(cpu->adr_stack, &cpu->stk_err);
+    cpu->ip = (size_t) StackPop(cpu->adr_stack, &cpu->stk_err);
 })
 
 DEF_CMD(IN, 17, NO_ARGS, {
